+++ package.json
@@ -1,0 +1,27 @@
+{
+  "name": "grocerieslistapp-yogigachinmath",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/pesto-students/grocerieslistapp-yogigachinmath.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/pesto-students/grocerieslistapp-yogigachinmath/issues"
+  },
+  "homepage": "https://github.com/pesto-students/grocerieslistapp-yogigachinmath#readme",
+  "dependencies": {
+    "webpack": "^5.11.0",
+    "webpack-cli": "^4.2.0"
+  },
+  "devDependencies": {
+  }
+}
