+++ public/index.html
@@ -7,13 +7,13 @@
 
     <!-- Bootstrap CSS -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-    <link rel="stylesheet" href=".././index.css"> 
+    <link rel="stylesheet" href="../styles/index.css"> 
     <title>Grocery App</title>
   </head>
   <body>
     <div class="container">
         <div class="loginForm input-group mb-3">
-            <form onsubmit="handleSubmit(event)">
+            <form onsubmit="handleLoginSubmit(event)">
                 <h1> UserDetails </h1>
                 <!-- <label for="username">UserName</label> -->
                 <input name = "username" id = "username" placeholder="Enter your Name" required/>
@@ -43,6 +43,6 @@
     <!-- <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script> -->
-    <script src=".././index.js"></script>
+    <script src="../dist/bundle.js"></script>
   </body>
 </html>
